[Refactor/AdminPage] NewAdmin을 Admin으로 이동. Page 파일들은 모두  pages 디렉토리 안에만 있도록 변경
