+++ src/pages/AdminCategoryPage.jsx
@@ -1,0 +1,5 @@
+import AdminCategory from "../components/AdminPage/AdminCategory";
+
+export default function AdminCategoryPage() {
+  return <AdminCategory />;
+}
