+++ src/pages/AdminMenuPage.jsx
@@ -1,0 +1,5 @@
+import AdminMenu from "../components/AdminPage/AdminMenu";
+
+export default function AdminMenuPage() {
+  return <AdminMenu />;
+}
