+++ src/pages/AdminOptionPage.jsx
@@ -1,0 +1,5 @@
+import AdminOption from "../components/AdminPage/AdminOption";
+
+export default function AdminOptionPage() {
+  return <AdminOption />;
+}
