+++ src/pages/AdminPage.jsx
@@ -1,7 +1,6 @@
 // import Admin from "../components/AdminPage/Admin";
-import NewAdmin from "../components/NewAdminPage/NewAdminPage";
+import Admin from "../components/AdminPage/Admin";
 
 export default function AdminPage() {
-  // return <Admin/>;
-  return <NewAdmin />;
+  return <Admin />;
 }
